refacto: split Service into two different classes
Service has been split into CompanyService and ComputerService.
Validator used to be an attribute of Service - it is now an independent
bean.
As all three of these classes are beans, this required to change the
way they are instantiated in CLI.java and in the Servlets.
The Servlets have thus become Spring Components - this breaks things
for now, but will make sense once they become Controllers.